Assignment2/vgg16/requirements.txt
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
iopath==0.1.10
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pillow==10.4.0
portalocker==2.10.1
ptflops==0.7.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
six==1.16.0
sympy==1.13.1
tabulate==0.9.0
termcolor==2.4.0
torch==2.3.1
torchinfo==1.8.0
torchprofile==0.0.4
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
yacs==0.1.8
